# Add coverage benchmark