--- a/pyperformance/data-files/benchmarks/bm_coverage/pyproject.toml
+++ b/pyperformance/data-files/benchmarks/bm_coverage/pyproject.toml
@@ -1,0 +1,12 @@
+[project]
+name = "pyperformance_bm_coverage"
+requires-python = ">=3.8"
+dependencies = [
+    "pyperf",
+    "coverage",
+]
+urls = {repository = "https://github.com/python/pyperformance"}
+dynamic = ["version"]
+
+[tool.pyperformance]
+name = "coverage"