Add a coverage benchmark to help benchmarking and avoid regressions such as https://github.com/python/cpython/issues/93516.