@kumaraditya303 2022-06-06
> Have you actually tried this before and after the regression you mentioned as example?

@gvanrossum: The regression I linked above [^1] is not caused by a single change but rather accumulation of multiple slowdown caused by multiple changes of which the most significant ones being https://github.com/python/cpython/pull/91666 and https://github.com/python/cpython/pull/27077. Since there is no benchmark in pyperformance to exercise coverage performance, the slowdown was unnoticed. Once this benchmark is added, it will help us pinpoint which commits caused the most slowdown with a simple `git bisect` between `3.10` and `main` branches.

[^1]: https://github.com/python/pyperformance/pull/213#issue-1261503016