@gvanrossum 2022-06-06
Nevertheless it would be interesting to see the numbers of running this BM on the state before and after said commits to demonstrate its usefulness.