@kumaraditya303 2022-06-06
@gvanrossum  Benchmarking before and after https://github.com/python/cpython/commit/944fffee8916cb94321fa33cd3a43f4108717746 commit

Results:

| Benchmark | before | after                |
|-----------|:------:|:--------------------:|
| coverage  | 161 ms | 213 ms: 1.32x slower |