@ericsnowcurrently 2022-06-06
> Question for @ericsnowcurrently -- Do you agree this is useful?

Yeah, it's a helpful indicator of impact.